各コントローラー　本人以外のアクセスについてリダイレクト

+++ app/controllers/reactions_controller.rb
@@ -1,4 +1,5 @@
 class ReactionsController < ApplicationController
+
   def create
     reaction = Reaction.create(reaction_params)
     redirect_to meeting_path(reaction.meeting.id)
@@ -6,6 +7,7 @@ class ReactionsController < ApplicationController
   
   def destroy
     reaction = Reaction.find(params[:id])
+    redirect_to root_path if current_user.id != reaction.user_id
     reaction.destroy
     redirect_to meeting_path(reaction.meeting.id)
   end
